requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
click>=8.1.7
httpx>=0.27.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
PyYAML>=6.0.1
pytest>=8.2.0
pytest-asyncio>=0.23.7
